@@ -1,4 +1,4 @@
-class PyrGeoFunc < ActiveRecord::Migration
+class PyrGeoFunc < ActiveRecord::Migration[5.1]
   def up
     execute <<CMD
 -- Haversine Formula based pyr_geo_distance in miles (constant is diameter of Earth in miles)
